pyproject.toml
[tool.poetry]
name = "fastnewsapi"
version = "0.1.0"
description = ""
authors = ["Roy Johnson <roy0219johnson@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.110.2"}
httpx = "^0.27.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
